리팩토링: 재사용 가능한 ImgBox, TextBox, Favorites, EditBtn 개별 component화
